caso02-examen/requirements.txt
Flask==3.0.0
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2
gspread==5.12.0
google-auth==2.23.3
